various bugs fixed

+++ arcor2_AREditor/Assets/BASE/Scripts/models/Orientation.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json;
 namespace IO.Swagger.Model {
 
   /// <summary>
-  /// Orientation(x:float, y:float, z:float, w:float)
+  /// Model of orientation.
   /// </summary>
   [DataContract]
   public class Orientation {
@@ -18,7 +18,7 @@ namespace IO.Swagger.Model {
     /// <value>Gets or Sets W</value>
     [DataMember(Name="w", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "w")]
-    public double? W { get; set; }
+    public float? W { get; set; }
 
     /// <summary>
     /// Gets or Sets X
@@ -26,7 +26,7 @@ namespace IO.Swagger.Model {
     /// <value>Gets or Sets X</value>
     [DataMember(Name="x", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "x")]
-    public double? X { get; set; }
+    public float? X { get; set; }
 
     /// <summary>
     /// Gets or Sets Y
@@ -34,7 +34,7 @@ namespace IO.Swagger.Model {
     /// <value>Gets or Sets Y</value>
     [DataMember(Name="y", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "y")]
-    public double? Y { get; set; }
+    public float? Y { get; set; }
 
     /// <summary>
     /// Gets or Sets Z
@@ -42,7 +42,7 @@ namespace IO.Swagger.Model {
     /// <value>Gets or Sets Z</value>
     [DataMember(Name="z", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "z")]
-    public double? Z { get; set; }
+    public float? Z { get; set; }
 
 
     /// <summary>
